Add more modulization, protect array data
